backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.8.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt

# For services
openai==1.12.0
pyyaml==6.0.1

# For FastAPI utilities / testing (already present or good to have)
python-multipart==0.0.6
requests==2.32.3

# Testing (already present)
pytest==7.4.3
httpx==0.25.1
